requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
email-validator
httpx
